+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/NewEventWorkspace/NewEventWorkspace.component.js
@@ -11,9 +11,9 @@ import { ConfirmDialog } from '../../../Dialogs/ConfirmDialog.component';
 import { Widget } from '../../../Widget';
 import { WidgetStageHeader } from './WidgetStageHeader';
 import { WidgetEventSchedule } from '../../../WidgetEventSchedule';
+import { addEnrollmentEventPageActionTypes } from '../enrollmentAddEventPage.actions';
 import type { Props } from './newEventWorkspace.types';
 import { useLocationQuery } from '../../../../utils/routing';
-
 
 const styles = () => ({
     innerWrapper: {
@@ -28,6 +28,8 @@ const NewEventWorkspacePlain = ({
     teiId,
     enrollmentId,
     dataEntryHasChanges,
+    onCancel,
+    onSave,
     classes,
     ...passOnProps
 }: Props) => {
@@ -82,6 +84,10 @@ const NewEventWorkspacePlain = ({
                         orgUnitId={orgUnitId}
                         teiId={teiId}
                         enrollmentId={enrollmentId}
+                        onSaveSuccessActionType={addEnrollmentEventPageActionTypes.EVENT_SAVE_SUCCESS}
+                        onSaveErrorActionType={addEnrollmentEventPageActionTypes.EVENT_SAVE_ERROR}
+                        onSave={onSave}
+                        onCancel={onCancel}
                         {...passOnProps}
                     />}
                     {mode === tabMode.SCHEDULE && <WidgetEventSchedule
@@ -93,6 +99,10 @@ const NewEventWorkspacePlain = ({
                         enrollmentId={enrollmentId}
                         enrollmentDate={enrollmentDate}
                         incidentDate={incidentDate}
+                        onSaveSuccessActionType={addEnrollmentEventPageActionTypes.EVENT_SCHEDULE_SUCCESS}
+                        onSaveErrorActionType={addEnrollmentEventPageActionTypes.EVENT_SCHEDULE_ERROR}
+                        onSave={onSave}
+                        onCancel={onCancel}
                     />}
                 </div>
             </Widget>