+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/enrollmentAddEventPage.actions.js
@@ -3,5 +3,7 @@
 export const addEnrollmentEventPageActionTypes = {
     EVENT_SAVE_SUCCESS: 'NewEnrollmentEventPage.saveEventSuccess',
     EVENT_SAVE_ERROR: 'NewEnrollmentEventPage.saveEventError',
+    EVENT_SCHEDULE_SUCCESS: 'ScheduleEvent.ScheduleEventSuccess',
+    EVENT_SCHEDULE_ERROR: 'ScheduleEvent.ScheduleEventError',
 };